+++ services/web/server/src/simcore_service_webserver/db.py
@@ -19,7 +19,8 @@ from simcore_postgres_database.utils_aiopg import (
 )
 from tenacity import retry
 
-from .db_config import CONFIG_SECTION_NAME, assert_valid_config
+from .db_config import CONFIG_SECTION_NAME
+from .db_settings import assert_valid_config
 
 THIS_MODULE_NAME = __name__.split(".")[-1]
 THIS_SERVICE_NAME = "postgres"
